lambda/notification_api/requirements.txt
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
